### 2019-05-17 #1457 @whedon
[PRE REVIEW]: A reproducible Snakemake pipeline to analyse Illumina paired-end data from ChiP-Seq experiments
**Submitting author:** @mgalland (<a href="http://orcid.org/0000-0003-2161-8689">marc galland</a>)
**Repository:** <a href="https://github.com/KoesGroup/Snakemake_ChIPseq_PE" target ="_blank">https://github.com/KoesGroup/Snakemake_ChIPseq_PE</a>
**Version:** v1.2.0
