### 2019-05-18 c21407e
add ref to JBrowse
