+++ bibliography.bib
@@ -182,3 +182,30 @@ note={Gruning, Bjorn Chilton, John Koster, Johannes Dale, Ryan Soranzo, Nicola v
 doi={10.1016/j.cels.2018.03.014},
 url={http://www.ncbi.nlm.nih.gov/pubmed/29953862}
 }
+
+@Article{Buels2016,
+author="Buels, Robert
+and Yao, Eric
+and Diesh, Colin M.
+and Hayes, Richard D.
+and Munoz-Torres, Monica
+and Helt, Gregg
+and Goodstein, David M.
+and Elsik, Christine G.
+and Lewis, Suzanna E.
+and Stein, Lincoln
+and Holmes, Ian H.",
+title="JBrowse: a dynamic web platform for genome visualization and analysis",
+journal="Genome Biology",
+year="2016",
+month="Apr",
+day="12",
+volume="17",
+number="1",
+pages="66",
+abstract="JBrowse is a fast and full-featured genome browser built with JavaScript and HTML5. It is easily embedded into websites or apps but can also be served as a standalone web page.",
+issn="1474-760X",
+doi="10.1186/s13059-016-0924-1",
+url="https://doi.org/10.1186/s13059-016-0924-1"
+}
+
