### 2019-05-18 #1457 @arfon
:wave: @mgalland - it looks like your BibTeX file is missing a comma. Could you please fix the file and ask Whedon to compile the paper again with `@whedon generate pdf`
