### 2019-05-18 #1457 @arfon
@mgymrek - would you be able to edit this submission for JOSS?
