+++ bibliography.bib
@@ -116,7 +116,7 @@ AVAILABILITY: http://samtools.sourceforge.net.},
 
 @article{Anaconda:2016,
  title = Anaconda Software Distribution,
- journal={Computer software. Vers. 2-2.4.0}.
+ journal={Computer software. Vers. 2-2.4.0},
  author={Anaconda},
  year={2016}
 }
@@ -208,4 +208,3 @@ issn="1474-760X",
 doi="10.1186/s13059-016-0924-1",
 url="https://doi.org/10.1186/s13059-016-0924-1"
 }
-
