### 2019-05-18 #1457 @arfon
@mgalland looks like L118 needs changing to `title = {Anaconda Software Distribution},`

Also, when I compile locally, Pandoc complains about the following issues with your citations:

```
pandoc-citeproc: reference Gruning:2018 not found
pandoc-citeproc: reference Ramiez:2016 not found
pandoc-citeproc: reference Buels:2016 not found
```
