+++ bibliography.bib
@@ -115,7 +115,7 @@ AVAILABILITY: http://samtools.sourceforge.net.},
 }
 
 @article{Anaconda:2016,
- title = Anaconda Software Distribution,
+ title = {Anaconda Software Distribution},
  journal={Computer software. Vers. 2-2.4.0},
  author={Anaconda},
  year={2016}
@@ -183,7 +183,7 @@ doi={10.1016/j.cels.2018.03.014},
 url={http://www.ncbi.nlm.nih.gov/pubmed/29953862}
 }
 
-@Article{Buels2016,
+@Article{Buels:2016,
 author="Buels, Robert
 and Yao, Eric
 and Diesh, Colin M.
