### 2019-05-20 #1457 @lpantano
@danielskatz, it seems a good fit, we'll do it.

@mgallan, can you try to fix the figures. They are not centered. We'll find the editor that could handle this meanwhile.
