### 2019-05-20 #1457 @brainstorm
👋  @vladsaveliev, would you like to review this pipeline for JOSS?
