### 2019-05-21 #1457 @brainstorm
@vladsaveliev Usually the reviewer checklist is fairly straightforward, the total time spent on the actual review, in my personal experience, is quite short (20-30min of codebase/tests/docs review), since you know your way through Snakemake, that could potentially be even less.
