### 2019-05-21 #1457 @brainstorm
@vladsaveliev https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
