### 2019-05-21 #1465 @vladsaveliev
**Review**

Snakemake_ChIPseq_PE is a solid Snakemake workflow, designed very much in Snakemake spirit, and makes use of its main features like conda or signularity-based runs. Self-contained assuming you have conda executable installed. Python code is clean, follows PEP guidelines an thus easy to read.

The tool is well documented, pretty clear instructions how to install and test it. One small suggestion is to remove ~/ in this command so it would work assuming a user cloned the repo and changed into it:

```
conda env create --name chipseq --file ~/envs/global_env.yaml
```

Feature-wise, one suggestion is the ability to start from BAM files as well. Often you either have an external pre-aligned BAMs, or want to do the heavy lifting with another software and use the workflow just for reporting.

Tests complete without any issues and the output corresponds to the description.

All good from my side.
