### 2019-05-21 #1465 @brainstorm
Vlad, Thanks a ton for this! @mgalland, could you please apply the fixes/suggestions and generate a preliminary DOI via Zenodo?
