### 2019-05-21 #1465 @brainstorm
@mgalland, before you put it on Zenodo, please make sure that:

1) The paper figures render well, they seem a bit off.
2) There's better spacing on the enumerations, paragraphs and overall text structure.

Other than that, LGTM.
