### 2019-05-21 #1465 @vladsaveliev
Have to clarify that I don't think that running from BAMs is a requirement to accept the article. Just a suggestion. Can go as is.

Agree about the figure rendering though, images look cropped by the right edge, especially the DAG plot.
