### 2019-05-21 #1465 @mgalland
Yes, running from BAMs would be nice but, in practice, I don't really know how to implement this...
Do you think the figures are cropped because their size is too big? Jihed will look into that tonight (Amsterdam time!).
