### 2019-05-21 #1465 @vladsaveliev
@mgalland Snakemake is just a Python extension, so you can use `if`-`else` statements whenever you want, e.g. like that:
https://github.com/umccr/10x/blob/master/ema/Snakefile#L49-L75
https://github.com/vladsaveliev/vcf_stuff/blob/master/vcf_stuff/eval_vcf/vcf.smk#L137-L156
Hope this is helpful
