### 2019-05-22 8cfbe24
Added line breaks before figures
