+++ paper.md
@@ -42,11 +42,14 @@ This pipeline also provides a small subsample of sequencing reads generated from
 
 
 # Figures
-- **Figure 1**: A Directed Acyclic Graph (DAG) of the Snakemake ChIP-seq PE pipeline. This graph has been produced with the command: `snakemake --rulegraph |dot -Tpng > dag.png`
+- **Figure 1**: A Directed Acyclic Graph (DAG) of the Snakemake ChIP-seq PE pipeline. This graph has been produced with the command: `snakemake --rulegraph |dot -Tpng > dag.png`.    
 ![Directed Acyclic Graph of rules](dag.png)
-- **Figure 2**: ChIP-seq tracks generated using the pipeline and visualized using JBrowse [@Buels:2016] ![Tracks](genome_viewer.png).
-- **Figure 3**: A pearson correlation plot generated by the pipeline using Deeptools to control for the quality of the experiment ![Correlation plot](correlationplot.png).
-- **Figure 4**: A profile plot showing the distribution of the reads over a selected genomic feature, here the genes ![Profile plot](profileplot.png).
+- **Figure 2**: ChIP-seq tracks generated using the pipeline and visualized using JBrowse [@Buels:2016].   
+![Tracks](genome_viewer.png).
+- **Figure 3**: A pearson correlation plot generated by the pipeline using Deeptools to control for the quality of the experiment.   
+![Correlation plot](correlationplot.png).  
+- **Figure 4**: A profile plot showing the distribution of the reads over a selected genomic feature, here genes are displayed.  
+![Profile plot](profileplot.png).
 
 # Acknowledgements
 We acknowledge contributions from Ming Tang and Johannes Koster for their inspired scripts. We would also like to thank the group of RNA biology and applied bioinformatics of the Swammerdam institute for Life sciences for providing the computational ressources, especially Wim de Leeuw and Han Rauwerda.
