### 2019-05-22 #1465 @mgalland
@brainstorm Figures seem fine now. Could you be more specific regarding the text formatting? Where would you like to see modifications? Thanks.
