### 2019-05-23 #1465 @brainstorm
Thanks @mgalland, looks better now. I was referring to the enumeration after "The outputs delivered with the pipeline are:"... an few extra newlines specifically there and better separation of paragraphs could improve readability. It is now a bit too compact imho.

After this is addressed, feel free to deposit on Zenodo. Cheers!
