### 2019-06-06 #1465 @brainstorm
@mgalland I see, my point was that spacing were similar to the GH-rendered markdown version https://github.com/KoesGroup/Snakemake_ChIPseq_PE/blob/master/paper.md, but there might be other tricks to display it well on @arfon's JOSS renderer.

Nevermind, looks good anyway, it was just a minor spacing detail. Looks good to me, could you generate a Zenodo DOI so that it can be archived?

Thanks everyone!
